# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Autoquímicos - Sistema de Gestión de Salas
  description: |
    API REST para el sistema de gestión de salas de conferencias de Autoquímicos S.A. de C.V.
    
    ## Características
    - Gestión de usuarios con roles (Administrador, Aprobador, Organizador, Recepción)
    - Reservas de salas con flujo de aprobación
    - Sistema de coffee break automático
    - Check-in y control de asistencia
    - Auditoría completa de acciones
    - Reportes y estadísticas
    
    ## Autenticación
    La API utiliza JWT (JSON Web Tokens) para autenticación. Incluye el token en el header:
    ```
    Authorization: Bearer <token>
    ```
    
  version: 1.0.0
  contact:
    name: Autoquímicos S.A. de C.V.
    email: admin@autoquimicos.com
  license:
    name: Propietario
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo
  - url: https://api.autoquimicos.com/api
    description: Servidor de producción
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Reservation:
      type: object
      properties:
        id_reserva:
          type: integer
          example: 1
        titulo:
          type: string
          example: "Reunión de ventas"
        asistentes:
          type: integer
          example: 8
        fecha:
          type: string
          format: date
          example: "2024-12-10"
        hora_inicio:
          type: string
          format: time
          example: "09:00:00"
        hora_fin:
          type: string
          format: time
          example: "11:00:00"
        require_coffee:
          type: boolean
          example: true
        status_name:
          type: string
          enum: [pendiente, aprobada, rechazada, cancelada, completada]
        organizer_name:
          type: string
          example: "Juan Pérez"
        room_name:
          type: string
          example: "Sala Ejecutiva"
...
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operación exitosa"
paths:
...
  /reservations:
    get:
      tags:
        - Reservas
      summary: Listar reservas
      description: Obtiene reservas con filtros opcionales
      security:
        - BearerAuth: []
      parameters:
        - name: roomId
          in: query
          schema:
            type: integer
          example: 1
        - name: status
          in: query
          schema:
            type: string
            enum: [pendiente, aprobada, rechazada, cancelada, completada]
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Lista de reservas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                    example: 5
                  reservations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Reservation'

    post:
      tags:
        - Reservas
      summary: Crear reserva
      description: Crea una nueva reserva de sala
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - idSala
                - titulo
                - asistentes
                - fecha
                - horaInicio
                - horaFin
              properties:
                idSala:
                  type: integer
                  example: 1
                titulo:
                  type: string
                  example: "Reunión de ventas"
                asistentes:
                  type: integer
                  example: 8
                fecha:
                  type: string
                  format: date
                  example: "2024-12-10"
                horaInicio:
                  type: string
                  format: time
                  example: "09:00:00"
                horaFin:
                  type: string
                  format: time
                  example: "11:00:00"
                requireCoffee:
                  type: boolean
                  example: true
      responses:
        '201':
          description: Reserva creada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Reserva creada exitosamente"
                  reservation:
                    $ref: '#/components/schemas/Reservation'
  /reservations/{id}/approve:
    post:
      tags:
        - Reservas
      summary: Aprobar reserva
      description: Aprueba una reserva pendiente (solo Aprobadores)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Reserva aprobada
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Reserva aprobada exitosamente"
                  reservation:
                    $ref: '#/components/schemas/Reservation'
  /reservations/{id}/reject:
    post:
      tags:
        - Reservas
      summary: Rechazar reserva
      description: Rechaza una reserva pendiente (solo Aprobadores)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  example: "Conflicto de horario con otra reunión"
      responses:
        '200':
          description: Reserva rechazada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
  /reservations/calendar/view:
    get:
      tags:
        - Reservas
      summary: Vista de calendario
      description: Obtiene reservas para vista de calendario
      security:
        - BearerAuth: []
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2024-12-01"
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2024-12-31"
        - name: roomId
          in: query
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Eventos del calendario
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Reservation'
...